Adding setup to enable flake8 and isort.
