+++ setup.cfg
@@ -1,0 +1,7 @@
+[flake8]
+max-complexity = 10
+max-line-length = 120
+exclude =
+    .git,
+    .pytest*,
+    .*_cache
