+++ tests/test_functions.py
@@ -1,5 +1,3 @@
-import pytest
-
 from utils import functions
