+++ tests/test_player.py
@@ -1,0 +1,7 @@
+from utils.player import DotaPlayerStats
+
+
+class TestDotaPlayerStats:
+    def test_basic(self):
+        tony = DotaPlayerStats(11)
+        assert tony.open_id == 11
